Move Admin API client to separate package
